Speed-limit CIFAR-10 experiment
The experiment is currently set to run for about two weeks on a single 24GB GPU. 

To run the experiment simply use ./runner.sh

If more GPUs are available, the runner.sh script can be easily split (the easiest is to split it in half between the two default learning rate and the 1e-7 learning rate). It also lists the approximate memory requirements of each individual run.